# schememitra/packages/contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "SchemeMitra Core API",
    "description": "Deterministic eligibility, repayment maths, partner routing, documents, voice and application tracking for marginalized entrepreneurs. Money is integer paise; rates are bps.",
    "version": "1.0.0"
  },
  "paths": {
...
    "/v1/applications": {
      "get": {
        "tags": [
          "applications"
        ],
        "summary": "List Mine",
        "operationId": "list_mine_v1_applications_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ApplicationOut"
                  },
                  "type": "array",
                  "title": "Response List Mine V1 Applications Get"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "applications"
        ],
        "summary": "Create Or Update",
        "description": "Idempotent on client_uuid: offline clients can retry safely.",
        "operationId": "create_or_update_v1_applications_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DraftIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApplicationOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/applications/{application_id}": {
      "get": {
        "tags": [
          "applications"
        ],
        "summary": "Get Application",
        "operationId": "get_application_v1_applications__application_id__get",
        "parameters": [
          {
            "name": "application_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Application Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApplicationOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/applications/{application_id}/readiness": {
      "post": {
        "tags": [
          "applications"
        ],
        "summary": "Readiness",
        "operationId": "readiness_v1_applications__application_id__readiness_post",
        "parameters": [
          {
            "name": "application_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Application Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReadinessOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/applications/{application_id}/consent": {
      "post": {
        "tags": [
          "applications"
        ],
        "summary": "Consent",
        "operationId": "consent_v1_applications__application_id__consent_post",
        "parameters": [
          {
            "name": "application_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Application Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConsentIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Consent V1 Applications  Application Id  Consent Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/applications/{application_id}/submit": {
      "post": {
        "tags": [
          "applications"
        ],
        "summary": "Submit",
        "operationId": "submit_v1_applications__application_id__submit_post",
        "parameters": [
          {
            "name": "application_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Application Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SubmitOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/applications/{application_id}/resubmit": {
      "post": {
        "tags": [
          "applications"
        ],
        "summary": "Resubmit",
        "operationId": "resubmit_v1_applications__application_id__resubmit_post",
        "parameters": [
          {
            "name": "application_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Application Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApplicationOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/applications/{application_id}/qr.png": {
      "get": {
        "tags": [
          "applications"
        ],
        "summary": "Qr Png",
        "operationId": "qr_png_v1_applications__application_id__qr_png_get",
        "parameters": [
          {
            "name": "application_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Application Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/applications/{application_id}/loan-file.pdf": {
      "get": {
        "tags": [
          "applications"
        ],
        "summary": "Loan File",
        "operationId": "loan_file_v1_applications__application_id__loan_file_pdf_get",
        "parameters": [
          {
            "name": "application_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Application Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/applications/{application_id}/sanction-letter.pdf": {
      "get": {
        "tags": [
          "applications"
        ],
        "summary": "Sanction Letter",
        "description": "Issued by the lender on approval; readable by the applicant side and that lender.",
        "operationId": "sanction_letter_v1_applications__application_id__sanction_letter_pdf_get",
        "parameters": [
          {
            "name": "application_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Application Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/applications/{application_id}/erase": {
      "post": {
        "tags": [
          "applications"
        ],
        "summary": "Erase",
        "description": "Withdraw consent and delete personal data (DPDP Act 2023, right to erasure).",
        "operationId": "erase_v1_applications__application_id__erase_post",
        "parameters": [
          {
            "name": "application_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Application Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "boolean"
                  },
                  "title": "Response Erase V1 Applications  Application Id  Erase Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/track/{tracking_id}": {
      "get": {
        "tags": [
          "applications"
        ],
        "summary": "Track",
        "description": "Public status by tracking ID (no personal data), also used by SMS STATUS and IVR.",
        "operationId": "track_v1_track__tracking_id__get",
        "parameters": [
          {
            "name": "tracking_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 10,
              "maxLength": 24,
              "title": "Tracking Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TrackOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/applications/{application_id}/consents": {
      "get": {
        "tags": [
          "applications"
        ],
        "summary": "Consents",
        "operationId": "consents_v1_applications__application_id__consents_get",
        "parameters": [
          {
            "name": "application_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Application Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "additionalProperties": true
                  },
                  "title": "Response Consents V1 Applications  Application Id  Consents Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ApplicationOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "tracking_id": {
            "type": "string",
            "title": "Tracking Id"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "scheme_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Scheme Code"
          },
          "partner_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Partner Id"
          },
          "partner_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Partner Name"
          },
          "lang": {
            "type": "string",
            "title": "Lang"
          },
          "submitted_via": {
            "type": "string",
            "title": "Submitted Via"
          },
          "completeness_score": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completeness Score"
          },
          "eligibility_status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Eligibility Status"
          },
          "plan": {
            "additionalProperties": true,
            "type": "object",
            "title": "Plan"
          },
          "finance": {
            "additionalProperties": true,
            "type": "object",
            "title": "Finance"
          },
          "requested_documents": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Requested Documents"
          },
          "rejection_reason_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rejection Reason Code"
          },
          "timeline": {
            "items": {
              "$ref": "#/components/schemas/TimelineItem"
            },
            "type": "array",
            "title": "Timeline"
          },
          "documents": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Documents"
          },
          "consent_recorded": {
            "type": "boolean",
            "title": "Consent Recorded"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          },
          "submitted_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Submitted At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "tracking_id",
          "status",
          "scheme_code",
          "partner_id",
          "partner_name",
          "lang",
          "submitted_via",
          "completeness_score",
          "eligibility_status",
          "plan",
          "finance",
          "requested_documents",
          "rejection_reason_code",
          "timeline",
          "documents",
          "consent_recorded",
          "created_at",
          "submitted_at"
        ],
        "title": "ApplicationOut"
      },
...
      "ConsentIn": {
        "properties": {
          "method": {
            "type": "string",
            "enum": [
              "otp",
              "thumb",
              "voice"
            ],
            "title": "Method"
          },
          "language": {
            "type": "string",
            "maxLength": 8,
            "title": "Language"
          },
          "text_version": {
            "type": "string",
            "maxLength": 32,
            "title": "Text Version",
            "default": "2026-09-v1"
          },
          "scope": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Scope",
            "default": [
              "profile",
              "documents",
              "share_partner"
            ]
          },
          "evidence": {
            "additionalProperties": true,
            "type": "object",
            "title": "Evidence",
            "default": {}
          }
        },
        "type": "object",
        "required": [
          "method",
          "language"
        ],
        "title": "ConsentIn"
      },
...
      "DraftIn": {
        "properties": {
          "client_uuid": {
            "type": "string",
            "maxLength": 64,
            "minLength": 8,
            "title": "Client Uuid"
          },
          "lang": {
            "type": "string",
            "maxLength": 8,
            "title": "Lang",
            "default": "en"
          },
          "personal": {
            "$ref": "#/components/schemas/PersonalIn",
            "default": {}
          },
          "facts": {
            "$ref": "#/components/schemas/ApplicantFacts"
          },
          "scheme_code": {
            "type": "string",
            "maxLength": 48,
            "title": "Scheme Code"
          },
          "plan": {
            "$ref": "#/components/schemas/LoanPlanIn"
          },
          "partner_id": {
            "type": "string",
            "title": "Partner Id"
          },
          "submitted_via": {
            "type": "string",
            "enum": [
              "app",
              "pwa",
              "sms",
              "csc"
            ],
            "title": "Submitted Via",
            "default": "pwa"
          }
        },
        "type": "object",
        "required": [
          "client_uuid",
          "facts",
          "scheme_code",
          "plan",
          "partner_id"
        ],
        "title": "DraftIn"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PersonalIn": {
        "properties": {
          "full_name": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 120
              },
              {
                "type": "null"
              }
            ],
            "title": "Full Name"
          },
          "father_name": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 120
              },
              {
                "type": "null"
              }
            ],
            "title": "Father Name"
          },
          "dob": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dob"
          },
          "phone": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^[6-9]\\d{9}$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Phone"
          }
        },
        "type": "object",
        "title": "PersonalIn"
      },
...
      "ReadinessOut": {
        "properties": {
          "score": {
            "type": "integer",
            "title": "Score"
          },
          "ready": {
            "type": "boolean",
            "title": "Ready"
          },
          "threshold": {
            "type": "integer",
            "title": "Threshold"
          },
          "components": {
            "additionalProperties": {
              "type": "number"
            },
            "type": "object",
            "title": "Components"
          },
          "blockers": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Blockers"
          },
          "fixes": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Fixes"
          },
          "name_matches": {
            "additionalProperties": true,
            "type": "object",
            "title": "Name Matches"
          },
          "required_documents": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Required Documents"
          }
        },
        "type": "object",
        "required": [
          "score",
          "ready",
          "threshold",
          "components",
          "blockers",
          "fixes",
          "name_matches",
          "required_documents"
        ],
        "title": "ReadinessOut"
      },
...
      "SubmitOut": {
        "properties": {
          "tracking_id": {
            "type": "string",
            "title": "Tracking Id"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "qr_png_url": {
            "type": "string",
            "title": "Qr Png Url"
          },
          "loan_file_pdf_url": {
            "type": "string",
            "title": "Loan File Pdf Url"
          },
          "qr_token": {
            "type": "string",
            "title": "Qr Token"
          }
        },
        "type": "object",
        "required": [
          "tracking_id",
          "status",
          "qr_png_url",
          "loan_file_pdf_url",
          "qr_token"
        ],
        "title": "SubmitOut"
      },
...
      "TimelineItem": {
        "properties": {
          "status": {
            "type": "string",
            "title": "Status"
          },
          "at": {
            "type": "string",
            "title": "At"
          },
          "note_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note Key"
          },
          "params": {
            "additionalProperties": true,
            "type": "object",
            "title": "Params",
            "default": {}
          }
        },
        "type": "object",
        "required": [
          "status",
          "at"
        ],
        "title": "TimelineItem"
      },
...
    }
  }
}